Hi,
as I worked on MQTT client library[0], I've found forum post about cerealed's speed[1], so I tried to make my "just mqtt" serialization as fast as msgpack and trying that I found some ways to improve performance of cerealed in specific cases and decided to make a PR for it.

I took your provided benchmark [2] and made some tuning for strings and byte arrays.

The test went from (DMD debug):
Cerealed: 5 secs, 926 ms, 824 μs, and 5 hnsecs
MsgPack:  2 secs, 226 ms, 575 μs, and 3 hnsecs

To:
Cerealed: 2 secs, 667 ms, 188 μs, and 3 hnsecs
MsgPack:  2 secs, 159 ms, 598 μs, and 1 hnsec

Which I like, so I hope this helps somehow. There are probably some other possible ways to tune it a bit more for other cases too.

[0] https://github.com/chalucha/vibe-mqtt
[1] http://forum.dlang.org/post/kceyzhlcumjxlfwbhcfr@forum.dlang.org
[2] http://dpaste.dzfl.pl/17b0ed9c0204