@atilaneves 2015-07-29
Interesting, this was already along the lines of what I wanted to do. Unfortunately, the code doesn't even compile with this patch (Travis is acting up so I downloaded the patch, applied it manually and tried running the UTs). I'll do a commit in a few min that will allow you to run them youself with `dub test`. Please rebase your PR then and fix the issue. Thanks!