@atilaneves 2015-07-29
I had time so I took a look: the issue is on line 89 when casting `val` to `ubyte[]`. Sometimes val in that case is a range and not an array, so simply appending `.array` to val there makes it compile and successfully run the tests. I'll let you fix your PR then I can merge, so you can get the credit in the git history. Thanks!