@chalucha 2015-07-29
Hi, I've added the to array fix, it compiles fine, but at least for me one of the tests fails:
tests.encode.testEncodeAssocArray:
    tests/encode.d:117 - Expected: [0, 2, 0, 0, 0, 1, 0, 0, 0, 2, 0, 0, 0, 3, 0, 0, 0, 6]
    tests/encode.d:117 -      Got: [0, 2, 0, 0, 0, 3, 0, 0, 0, 6, 0, 0, 0, 1, 0, 0, 0, 2]

I didn't touch AAs serialization so I hope it isn't caused by this.