@atilaneves 2015-07-30
Weird, not only does it pass for me, as you can see from this page it passes on Travis as well. I'll merge it anyway. What's your OS / architecture / etc. ? A failing test is... not good. But I can't reproduce the issue.